feat: update CHANGELOG and package.json for version 1.6.0, add fetch-token command and project initialization flow
