+++ packages/cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@hiveops/cli",
-  "version": "1.3.0",
+  "version": "1.4.0",
   "main": "./lib/index.js",
   "types": "./lib/index.d.ts",
   "scripts": {
@@ -8,7 +8,6 @@
     "dev:watch": "nodemon src/main.ts",
     "fetch-token": "ts-node src/hive.ts vespa fetch-token",
     "initialize-project": "ts-node src/hive.ts vespa init --project-directory ../../../examples/vespa-typescript",
-    "initialize-project-v2": "ts-node src/hive.ts vespa init-v2 --project-directory ../../../examples/vespa-typescript",
     "generate-code": "ts-node src/hive.ts vespa generate-code",
     "apply-migration": "ts-node src/hive.ts vespa apply --project-directory ../../../examples/vespa-typescript",
     "clean": "rm -rf lib/",
@@ -56,7 +55,7 @@
   },
   "dependencies": {
     "@hiveops/core": "^1.6.0",
-    "@hiveops/node": "^1.4.0",
+    "@hiveops/node": "^1.6.0",
     "commander": "^13.1.0",
     "dotenv": "^16.5.0",
     "express": "^5.1.0",
