+++ packages/node/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@hiveops/node",
-  "version": "1.5.0",
+  "version": "1.6.0",
   "main": "./lib/index.js",
   "types": "./lib/index.d.ts",
   "scripts": {
